Refactor: replace old VisuallyHidden component with new
